requirements.txt
Django==4.2.*
djangorestframework==3.14.*
djoser==2.2.*
djangorestframework-simplejwt==5.2.*
psycopg2-binary==2.9.*
celery==5.2.*
redis==4.5.*
drf-yasg==1.21.*
python-dotenv==1.0.*
gunicorn==20.1.*
pytest==7.3.*
pytest-django==4.5.*
coverage==7.2.*
flake8==6.0.*
black==23.3.*
